Use timestamp locker in REST services.

+++ src/main/java/de/intevation/lada/rest/MesswertService.java
@@ -25,6 +25,9 @@ import javax.ws.rs.core.UriInfo;
 
 import org.apache.log4j.Logger;
 
+import de.intevation.lada.lock.LockConfig;
+import de.intevation.lada.lock.LockType;
+import de.intevation.lada.lock.ObjectLocker;
 import de.intevation.lada.model.land.LMesswert;
 import de.intevation.lada.util.annotation.AuthorizationConfig;
 import de.intevation.lada.util.annotation.RepositoryConfig;
@@ -48,6 +51,10 @@ public class MesswertService {
     @Inject
     @RepositoryConfig(type=RepositoryType.RW)
     private Repository defaultRepo;
+
+    @Inject
+    @LockConfig(type=LockType.TIMESTAMP)
+    private ObjectLocker lock;
 
     /* The authorization module.*/
     @Inject
@@ -144,6 +151,9 @@ public class MesswertService {
         ) {
             return new Response(false, 699, null);
         }
+        if (lock.isLocked(messwert)) {
+            return new Response(false, 697, null);
+        }
         Response response = defaultRepo.update(messwert, "land");
         Response updated = defaultRepo.getById(
             LMesswert.class,
@@ -175,6 +185,9 @@ public class MesswertService {
                 LMesswert.class)
         ) {
             return new Response(false, 699, null);
+        }
+        if (lock.isLocked(messwert)) {
+            return new Response(false, 697, null);
         }
         /* Delete the messwert object*/
         return defaultRepo.delete(messwertObj, "land");
